# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Video to Shorts Converter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  uploading: 'uploading',
  analyzing: 'analyzing',
  clipping: 'clipping',
  rendering: 'rendering',
  ready: 'ready',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export type JobPlatform = typeof JobPlatform[keyof typeof JobPlatform];


export const JobPlatform = {
  shorts: 'shorts',
  reels: 'reels',
  both: 'both',
} as const;

export interface Job {
  id: number;
  title: string;
  status: JobStatus;
  /** 0-100 */
  progress: number;
  platform: JobPlatform;
  /** Target clip duration in seconds (15, 30, or 60) */
  clipDuration: number;
  autoCaptions: boolean;
  /** @nullable */
  originalFilename?: string | null;
  /** @nullable */
  originalDurationSeconds?: number | null;
  /** @nullable */
  fileSizeMb?: number | null;
  /** @nullable */
  estimatedSecondsRemaining?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type JobInputPlatform = typeof JobInputPlatform[keyof typeof JobInputPlatform];


export const JobInputPlatform = {
  shorts: 'shorts',
  reels: 'reels',
  both: 'both',
} as const;

export type JobInputClipDuration = typeof JobInputClipDuration[keyof typeof JobInputClipDuration];


export const JobInputClipDuration = {
  NUMBER_15: 15,
  NUMBER_30: 30,
  NUMBER_60: 60,
} as const;

export interface JobInput {
  /** @minLength 1 */
  title: string;
  platform: JobInputPlatform;
  clipDuration: JobInputClipDuration;
  autoCaptions: boolean;
  originalFilename?: string;
  originalDurationSeconds?: number;
  fileSizeMb?: number;
}

export type JobProgressUpdateStatus = typeof JobProgressUpdateStatus[keyof typeof JobProgressUpdateStatus];


export const JobProgressUpdateStatus = {
  pending: 'pending',
  uploading: 'uploading',
  analyzing: 'analyzing',
  clipping: 'clipping',
  rendering: 'rendering',
  ready: 'ready',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface JobProgressUpdate {
  status?: JobProgressUpdateStatus;
  progress?: number;
  estimatedSecondsRemaining?: number;
  errorMessage?: string;
}

export type ClipPlatform = typeof ClipPlatform[keyof typeof ClipPlatform];


export const ClipPlatform = {
  shorts: 'shorts',
  reels: 'reels',
  both: 'both',
} as const;

export interface Clip {
  id: number;
  jobId: number;
  name: string;
  durationSeconds: number;
  startTimeSeconds: number;
  platform: ClipPlatform;
  /** @nullable */
  sceneLabel?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /**
     * AI-estimated viral potential 0-100
     * @nullable
     */
  viralScore?: number | null;
  createdAt: string;
}

export interface ClipUpdate {
  /** @minLength 1 */
  name?: string;
}

export interface Stats {
  totalJobs: number;
  completedJobs: number;
  processingJobs: number;
  totalClipsGenerated: number;
  totalVideoMinutesProcessed: number;
  averageClipsPerJob: number;
}