# lib/api-zod/src/generated/types/clip.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Video to Shorts Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ClipPlatform } from './clipPlatform';

export interface Clip {
  id: number;
  jobId: number;
  name: string;
  durationSeconds: number;
  startTimeSeconds: number;
  platform: ClipPlatform;
  /** @nullable */
  sceneLabel?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /**
     * AI-estimated viral potential 0-100
     * @nullable
     */
  viralScore?: number | null;
  createdAt: Date;
}