# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Video to Shorts Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { JobInputClipDuration } from './jobInputClipDuration';
import type { JobInputPlatform } from './jobInputPlatform';

export interface JobInput {
  /** @minLength 1 */
  title: string;
  platform: JobInputPlatform;
  clipDuration: JobInputClipDuration;
  autoCaptions: boolean;
  originalFilename?: string;
  originalDurationSeconds?: number;
  fileSizeMb?: number;
}