# lib/api-zod/src/generated/types/jobProgressUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Video to Shorts Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { JobProgressUpdateStatus } from './jobProgressUpdateStatus';

export interface JobProgressUpdate {
  status?: JobProgressUpdateStatus;
  progress?: number;
  estimatedSecondsRemaining?: number;
  errorMessage?: string;
}