# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Video to Shorts Converter API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalJobs: number;
  completedJobs: number;
  processingJobs: number;
  totalClipsGenerated: number;
  totalVideoMinutesProcessed: number;
  averageClipsPerJob: number;
}